requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
openai>=1.7.0
python-dotenv==1.0.0
requests==2.31.0
